requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
asgiref=3.3.0=py_0
blinker=1.4=py38_0
brotlipy=0.7.0=py38haf1e3a3_1000
ca-certificates=2020.10.14=0
certifi=2020.6.20=py38_0
cffi=1.14.3=py38hed5b41f_0
chardet=3.0.4=py38_1003
chembl_webresource_client=0.10.1=pyh9f0ad1d_0
cryptography=3.1.1=py38hddc9c9b_0
defusedxml=0.6.0=py_0
django=3.1.2=py_0
django-crispy-forms=1.8.1=py_0
djangorestframework=3.12.1=py_0
easydict=1.9=py_0
ecdsa=0.16.0=py_0
future=0.18.2=py38_1
gevent=20.9.0=py38haf1e3a3_0
gmp=6.1.2=hb37e062_1
greenlet=0.4.17=py38haf1e3a3_0
grequests=0.6.0=pyh9f0ad1d_0
idna=2.10=py_0
jinja2=2.11.2=py_0
krb5=1.17.1=hddcf347_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libpq=12.2=h051b688_0
markupsafe=1.1.1=py38h1de35cc_1
mypy-extensions=0.4.3=pypi_0
mysql-connector-c=6.1.11=hccea1a4_1
mysqlclient=2.0.1=py38hb1e8313_0
ncurses=6.2=h0a44026_1
oauthlib=3.1.0=py_0
openssl=1.1.1h=haf1e3a3_0
pip=20.2.4=py38_0
psycopg2=2.8.5=py38hddc9c9b_0
pycparser=2.20=py_2
pycryptodome=3.9.8=py38h9066e36_0
pyjwt=1.7.1=py38_0
pyopenssl=19.1.0=py_1
pysocks=1.7.1=py38_1
python=3.8.5=h26836e1_1
python-jose=3.2.0=py_0
python3-openid=3.1.0=py38_1001
pytz=2020.1=py_0
qwikidata=0.4.0=pypi_0
readline=8.0=h1de35cc_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=py_0
requests_cache=0.4.13=py_0
setuptools=50.3.0=py38h0dc7051_1
six=1.15.0=py_0
slacker=0.14.0=py_0
social-auth-app-django=4.0.0=pyh9f0ad1d_0
social-auth-core=3.3.0=pyh8c360ce_0
sqlite=3.33.0=hffcf06c_0
sqlparse=0.4.1=py_0
tk=8.6.10=hb0a8c7a_0
unidecode=1.1.1=py_0
urllib3=1.25.11=py_0
wheel=0.35.1=py_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
zope=1.0=py38_1
zope.event=4.5.0=py38_0
zope.interface=5.1.2=py38haf1e3a3_0
